Split models and books module
